gllcms: Use `get_or_insert_with` to solve `clippy::option_if_let_else`
The suggestion to use `map_or_else` here - or `unwrap_or_else` which
prevents a useless "identity" closure - is flawed as it does not allow a
secondary mutable borrow of the `state` lock to occur within the
closure.  `get_or_insert_with` neatly circumvents this problem with a
tidy pattern, no mutable assignments of `Some` values, and returns a
mutable value (not `Option`) as a result.